# openapi/orbithr-v1.yaml
openapi: 3.1.0
info:
  title: OrbitHR API
  version: 1.0.0
  description: Authenticated, tenant-isolated API for OrbitHR web and Android clients.
servers:
  - url: http://localhost:3001/api/v1
security:
  - bearerAuth: []
paths:
...
  /workflows/definitions/{id}/activate:
    post:
      summary: Activate a workflow definition version
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'200': {description: Definition activated}}
...
  /workflows/instances/{id}/actions:
    post:
      summary: Approve, reject, or comment on the active step
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'200': {description: Action applied}}
  /workflows/instances/{id}/withdraw:
    post:
      summary: Withdraw a pending request owned by the current user
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'200': {description: Request withdrawn}}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...