# server/v1/workflows.ts
import { Router, type Request, type RequestHandler, type Response } from 'express';
import type { Prisma, PrismaClient, WorkflowApproverType, WorkflowModule } from '@prisma/client';
import { z } from 'zod';
import { emitNotification } from './notifications.js';

type WorkflowRequest = Request & { auth?: { id: string; companyId: string; role: string; employeeId?: string; permissions: string[] }; requestId?: string };
const uuid = z.string().uuid();
const modules = ['LEAVE','EXPENSE','ATTENDANCE_CORRECTION','WFH','ON_DUTY','BUSINESS_TRAVEL','OVERTIME','SALARY_REVISION','PAYROLL','DOCUMENT','ADVANCE','GENERIC'] as const;
const approverTypes = ['USER','ROLE','REPORTING_MANAGER','DEPARTMENT_HEAD','FINANCE'] as const;

type WorkflowDb = PrismaClient | Prisma.TransactionClient;
async function adapterApprovers(prisma: WorkflowDb, companyId: string, requesterUserId: string, type: WorkflowApproverType, reference?: string | null) {
  if (type === 'USER') return reference && await prisma.user.findFirst({ where: { id: reference, companyId } }) ? [reference] : [];
  if (type === 'ROLE') return (await prisma.user.findMany({ where: { companyId, role: reference as never }, select: { id: true } })).map(user => user.id);
  if (type === 'FINANCE') return (await prisma.user.findMany({ where: { companyId, role: { in: ['PAYROLL_ADMIN', 'COMPANY_ADMIN'] } }, select: { id: true } })).map(user => user.id);
  const requester = await prisma.employee.findFirst({ where: { companyId, userId: requesterUserId }, select: { reportingManager: { select: { userId: true } }, department: { select: { headEmployeeId: true } } } });
  if (type === 'REPORTING_MANAGER') return requester?.reportingManager?.userId ? [requester.reportingManager.userId] : [];
  if (type === 'DEPARTMENT_HEAD' && requester?.department.headEmployeeId) { const head = await prisma.employee.findFirst({ where: { id: requester.department.headEmployeeId, companyId }, select: { userId: true } }); return head?.userId ? [head.userId] : []; }
  return [];
}

export async function startConfiguredWorkflow(prisma: WorkflowDb, input: { companyId: string; requesterUserId: string; module: WorkflowModule; subjectType: string; subjectId: string; title: string; summary?: string; payload?: Prisma.InputJsonValue }) {
  const definition = await prisma.workflowDefinition.findFirst({ where: { companyId: input.companyId, module: input.module, status: 'ACTIVE' }, include: { steps: { orderBy: { sequence: 'asc' } } }, orderBy: { version: 'desc' } });
  if (!definition) return null;
  const stepData = []; for (const step of definition.steps) { const approverUserIds = await adapterApprovers(prisma, input.companyId, input.requesterUserId, step.approverType, step.approverReference); if (approverUserIds.length < step.minimumApprovals) throw Object.assign(new Error(`Step ${step.sequence} does not have enough eligible approvers.`), { status: 409, code: 'APPROVER_UNRESOLVED' }); stepData.push({ stepId: step.id, sequence: step.sequence, status: step.sequence === 1 ? 'PENDING' as const : 'WAITING' as const, approverUserIds, minimumApprovals: step.minimumApprovals, startedAt: step.sequence === 1 ? new Date() : undefined, dueAt: step.sequence === 1 && step.slaHours ? new Date(Date.now() + step.slaHours * 3_600_000) : undefined }); }
  const instance = await prisma.workflowInstance.create({ data: { companyId: input.companyId, definitionId: definition.id, module: input.module, subjectType: input.subjectType, subjectId: input.subjectId, requesterUserId: input.requesterUserId, title: input.title, summary: input.summary, payload: input.payload, stepInstances: { create: stepData }, actions: { create: { actorUserId: input.requesterUserId, action: 'SUBMIT', comment: input.summary } } } });
  for (const userId of stepData[0]?.approverUserIds || []) await emitNotification(prisma, { companyId: input.companyId, userId, eventKey: 'WORKFLOW_APPROVAL_REQUIRED', title: 'Approval required', body: input.title, entityType: 'WorkflowInstance', entityId: instance.id, actionUrl: '/approvals' });
  return instance;
}

export function createWorkflowRouter(prisma: PrismaClient, authenticate: RequestHandler) {
  const router = Router();
  const ok = (res: Response, data: unknown, status = 200) => res.status(status).json({ data, meta: { requestId: (res.req as WorkflowRequest).requestId } });
  const fail = (res: Response, status: number, code: string, message: string) => res.status(status).json({ error: { code, message }, meta: { requestId: (res.req as WorkflowRequest).requestId } });
  const permitted = (permission: string): RequestHandler => (req: WorkflowRequest, res, next) => req.auth?.permissions.includes(permission) ? next() : fail(res, 403, 'FORBIDDEN', 'You do not have permission to perform this action.');
  const audit = (req: WorkflowRequest, action: string, details: string) => prisma.auditLog.create({ data: { companyId: req.auth!.companyId, userId: req.auth!.id, userName: req.auth!.id, userRole: req.auth!.role, action, category: 'WORKFLOW', details, ipAddress: req.ip || 'unknown' } });

  async function resolveApprovers(companyId: string, requesterUserId: string, type: WorkflowApproverType, reference?: string | null): Promise<string[]> {
    if (type === 'USER') return reference && await prisma.user.findFirst({ where: { id: reference, companyId } }) ? [reference] : [];
    if (type === 'ROLE') return (await prisma.user.findMany({ where: { companyId, role: reference as never }, select: { id: true } })).map(user => user.id);
    if (type === 'FINANCE') return (await prisma.user.findMany({ where: { companyId, role: { in: ['PAYROLL_ADMIN', 'COMPANY_ADMIN'] } }, select: { id: true } })).map(user => user.id);
    const requester = await prisma.employee.findFirst({ where: { companyId, userId: requesterUserId }, select: { reportingManager: { select: { userId: true } }, department: { select: { headEmployeeId: true } } } });
    if (type === 'REPORTING_MANAGER') return requester?.reportingManager?.userId ? [requester.reportingManager.userId] : [];
    if (type === 'DEPARTMENT_HEAD' && requester?.department.headEmployeeId) {
      const head = await prisma.employee.findFirst({ where: { id: requester.department.headEmployeeId, companyId }, select: { userId: true } }); return head?.userId ? [head.userId] : [];
    }
    return [];
  }

  router.use('/workflows', authenticate);
  router.get('/workflows/definitions', permitted('workflow.manage'), async (req: WorkflowRequest, res, next) => { try { return ok(res, await prisma.workflowDefinition.findMany({ where: { companyId: req.auth!.companyId }, include: { steps: { orderBy: { sequence: 'asc' } } }, orderBy: [{ code: 'asc' }, { version: 'desc' }] })); } catch (error) { next(error); } });
  router.post('/workflows/definitions', permitted('workflow.manage'), async (req: WorkflowRequest, res, next) => { try {
    const body = z.object({ module: z.enum(modules), name: z.string().trim().min(2).max(120), code: z.string().trim().min(2).max(40).transform(v => v.toUpperCase()), criteria: z.record(z.string(), z.unknown()).optional(), steps: z.array(z.object({ name: z.string().trim().min(2).max(120), approverType: z.enum(approverTypes), approverReference: z.string().trim().max(100).optional(), minimumApprovals: z.number().int().min(1).max(20).default(1), slaHours: z.number().int().min(1).max(8760).optional(), allowDelegation: z.boolean().default(true), conditions: z.record(z.string(), z.unknown()).optional() })).min(1).max(20) }).parse(req.body);
    const companyId = req.auth!.companyId; const latest = await prisma.workflowDefinition.findFirst({ where: { companyId, code: body.code }, orderBy: { version: 'desc' }, select: { version: true } });
    const value = await prisma.workflowDefinition.create({ data: { companyId, module: body.module, name: body.name, code: body.code, version: (latest?.version || 0) + 1, criteria: body.criteria as never, createdById: req.auth!.id, steps: { create: body.steps.map((step, index) => ({ ...step, sequence: index + 1, conditions: step.conditions as never })) } }, include: { steps: { orderBy: { sequence: 'asc' } } } });
    await audit(req, 'CREATE_WORKFLOW_DEFINITION', `${value.code} v${value.version} created.`); return ok(res, value, 201);
  } catch (error) { next(error); } });
  router.post('/workflows/definitions/:id/activate', permitted('workflow.manage'), async (req: WorkflowRequest, res, next) => { try {
    const definition = await prisma.workflowDefinition.findFirst({ where: { id: String(req.params.id), companyId: req.auth!.companyId } }); if (!definition) return fail(res, 404, 'WORKFLOW_NOT_FOUND', 'Workflow definition was not found.');
    await prisma.$transaction([prisma.workflowDefinition.updateMany({ where: { companyId: definition.companyId, code: definition.code, status: 'ACTIVE' }, data: { status: 'RETIRED' } }), prisma.workflowDefinition.update({ where: { id: definition.id }, data: { status: 'ACTIVE' } }), audit(req, 'ACTIVATE_WORKFLOW_DEFINITION', `${definition.code} v${definition.version} activated.`)]); return ok(res, { activated: true });
  } catch (error) { next(error); } });

  router.post('/workflows/submit', async (req: WorkflowRequest, res, next) => { try {
    const body = z.object({ module: z.enum(modules), subjectType: z.string().trim().min(2).max(80), subjectId: z.string().trim().min(1).max(120), title: z.string().trim().min(2).max(180), summary: z.string().trim().max(1000).optional(), payload: z.record(z.string(), z.unknown()).optional() }).parse(req.body); const companyId = req.auth!.companyId;
    const definition = await prisma.workflowDefinition.findFirst({ where: { companyId, module: body.module, status: 'ACTIVE' }, include: { steps: { orderBy: { sequence: 'asc' } } }, orderBy: { version: 'desc' } }); if (!definition) return fail(res, 409, 'WORKFLOW_NOT_CONFIGURED', `No active ${body.module} workflow is configured.`);
    const stepData = []; for (const step of definition.steps) { const approverUserIds = await resolveApprovers(companyId, req.auth!.id, step.approverType, step.approverReference); if (approverUserIds.length < step.minimumApprovals) return fail(res, 409, 'APPROVER_UNRESOLVED', `Step ${step.sequence} does not have enough eligible approvers.`); stepData.push({ stepId: step.id, sequence: step.sequence, status: step.sequence === 1 ? 'PENDING' as const : 'WAITING' as const, approverUserIds, minimumApprovals: step.minimumApprovals, startedAt: step.sequence === 1 ? new Date() : undefined, dueAt: step.sequence === 1 && step.slaHours ? new Date(Date.now() + step.slaHours * 3_600_000) : undefined }); }
    const value = await prisma.workflowInstance.create({ data: { companyId, definitionId: definition.id, module: body.module, subjectType: body.subjectType, subjectId: body.subjectId, requesterUserId: req.auth!.id, title: body.title, summary: body.summary, payload: body.payload as never, stepInstances: { create: stepData }, actions: { create: { actorUserId: req.auth!.id, action: 'SUBMIT', comment: body.summary } } }, include: { stepInstances: true, actions: true } }); await audit(req, 'SUBMIT_WORKFLOW', `${body.module} request ${value.id} submitted.`); return ok(res, value, 201);
  } catch (error) { next(error); } });

  router.get('/workflows/my-requests', async (req: WorkflowRequest, res, next) => { try { return ok(res, await prisma.workflowInstance.findMany({ where: { companyId: req.auth!.companyId, requesterUserId: req.auth!.id }, include: { stepInstances: { orderBy: { sequence: 'asc' } }, actions: { orderBy: { createdAt: 'asc' } } }, orderBy: { submittedAt: 'desc' } })); } catch (error) { next(error); } });
  router.get('/workflows/inbox', permitted('workflow.review'), async (req: WorkflowRequest, res, next) => { try {
    const now = new Date(); const delegated = await prisma.workflowDelegation.findMany({ where: { companyId: req.auth!.companyId, delegateUserId: req.auth!.id, active: true, startsAt: { lte: now }, endsAt: { gte: now } } }); const principals = [req.auth!.id, ...delegated.map(value => value.delegatorUserId)];
    const items = await prisma.workflowStepInstance.findMany({ where: { status: 'PENDING', approverUserIds: { hasSome: principals }, instance: { companyId: req.auth!.companyId, status: 'PENDING' } }, include: { instance: true, step: true }, orderBy: [{ dueAt: 'asc' }, { startedAt: 'asc' }] }); return ok(res, items);
  } catch (error) { next(error); } });

  router.post('/workflows/instances/:id/withdraw', async (req: WorkflowRequest, res, next) => { try {
    const body = z.object({ comment: z.string().trim().max(1000).optional() }).parse(req.body); const companyId = req.auth!.companyId;
    const instance = await prisma.workflowInstance.findFirst({ where: { id: String(req.params.id), companyId, requesterUserId: req.auth!.id, status: 'PENDING' } });
    if (!instance) return fail(res, 404, 'WORKFLOW_NOT_WITHDRAWABLE', 'A pending request owned by you was not found.');
    const operations = [prisma.workflowInstance.update({ where: { id: instance.id }, data: { status: 'WITHDRAWN', cancelledAt: new Date(), completedAt: new Date() } }), prisma.workflowStepInstance.updateMany({ where: { instanceId: instance.id, status: { in: ['WAITING', 'PENDING'] } }, data: { status: 'SKIPPED', completedAt: new Date() } }), prisma.workflowAction.create({ data: { instanceId: instance.id, actorUserId: req.auth!.id, action: 'WITHDRAW', comment: body.comment } }), audit(req, 'WITHDRAW_WORKFLOW', `Workflow ${instance.id} withdrawn.`)];
    if (instance.subjectType === 'LeaveRequest') operations.push(prisma.leaveRequest.updateMany({ where: { id: instance.subjectId, companyId }, data: { status: 'CANCELLED' } }) as never);
    if (instance.subjectType === 'EmployeeServiceRequest') operations.push(prisma.employeeServiceRequest.updateMany({ where: { id: instance.subjectId, companyId }, data: { status: 'WITHDRAWN' } }) as never);
    if (instance.subjectType === 'AttendanceRequest') operations.push(prisma.attendanceRequest.updateMany({ where: { id: instance.subjectId, companyId }, data: { status: 'CANCELLED' } }) as never);
    await prisma.$transaction(operations); return ok(res, { status: 'WITHDRAWN' });
  } catch (error) { next(error); } });

  router.post('/workflows/instances/:id/actions', permitted('workflow.review'), async (req: WorkflowRequest, res, next) => { try {
    const body = z.object({ action: z.enum(['APPROVE','REJECT','COMMENT']), comment: z.string().trim().max(1000).optional() }).parse(req.body); const companyId = req.auth!.companyId;
    const instance = await prisma.workflowInstance.findFirst({ where: { id: String(req.params.id), companyId }, include: { stepInstances: { include: { step: true }, orderBy: { sequence: 'asc' } } } }); if (!instance || instance.status !== 'PENDING') return fail(res, 404, 'WORKFLOW_INSTANCE_NOT_PENDING', 'Pending workflow instance was not found.'); const step = instance.stepInstances.find(value => value.status === 'PENDING'); if (!step) return fail(res, 409, 'WORKFLOW_STEP_MISSING', 'No active approval step exists.');
    let actingForUserId: string | undefined; if (!step.approverUserIds.includes(req.auth!.id)) { if (!step.step.allowDelegation) return fail(res, 403, 'DELEGATION_DISABLED', 'Delegation is disabled for this approval step.'); const delegation = await prisma.workflowDelegation.findFirst({ where: { companyId, delegateUserId: req.auth!.id, delegatorUserId: { in: step.approverUserIds }, active: true, startsAt: { lte: new Date() }, endsAt: { gte: new Date() }, OR: [{ module: null }, { module: instance.module }] } }); if (!delegation) return fail(res, 403, 'NOT_ASSIGNED_APPROVER', 'This approval is not assigned to you.'); actingForUserId = delegation.delegatorUserId; }
    if (body.action === 'COMMENT') { const value = await prisma.workflowAction.create({ data: { instanceId: instance.id, stepInstanceId: step.id, actorUserId: req.auth!.id, actingForUserId, action: 'COMMENT', comment: body.comment } }); return ok(res, value, 201); }
    if (body.action === 'REJECT') { const operations = [prisma.workflowStepInstance.update({ where: { id: step.id }, data: { status: 'REJECTED', completedAt: new Date() } }), prisma.workflowInstance.update({ where: { id: instance.id }, data: { status: 'REJECTED', completedAt: new Date() } }), prisma.workflowAction.create({ data: { instanceId: instance.id, stepInstanceId: step.id, actorUserId: req.auth!.id, actingForUserId, action: 'REJECT', comment: body.comment } }), audit(req, 'REJECT_WORKFLOW', `Workflow ${instance.id} rejected.`)]; if (instance.subjectType === 'LeaveRequest') operations.push(prisma.leaveRequest.updateMany({ where: { id: instance.subjectId, companyId }, data: { status: 'REJECTED', approvedById: req.auth!.id, approvedAt: new Date() } }) as never); if (instance.subjectType === 'ExpenseClaim') operations.push(prisma.expenseClaim.updateMany({ where: { id: instance.subjectId, companyId }, data: { status: 'REJECTED', approvedById: req.auth!.id, approvedAt: new Date() } }) as never); if(instance.subjectType==='EmployeeServiceRequest')operations.push(prisma.employeeServiceRequest.updateMany({where:{id:instance.subjectId,companyId},data:{status:'REJECTED'}})as never); await prisma.$transaction(operations); return ok(res, { status: 'REJECTED' }); }
    const approvalPrincipal = actingForUserId || req.auth!.id; const duplicateApproval = await prisma.workflowAction.findFirst({ where: { stepInstanceId: step.id, action: 'APPROVE', OR: [{ actorUserId: approvalPrincipal }, { actingForUserId: approvalPrincipal }] } }); if (duplicateApproval) return fail(res, 409, 'ALREADY_APPROVED', 'This approver has already approved this step.');
    const approvals = step.approvals + 1; if (approvals < step.minimumApprovals) { await prisma.$transaction([prisma.workflowStepInstance.update({ where: { id: step.id }, data: { approvals } }), prisma.workflowAction.create({ data: { instanceId: instance.id, stepInstanceId: step.id, actorUserId: req.auth!.id, actingForUserId, action: 'APPROVE', comment: body.comment } })]); return ok(res, { status: 'PENDING', approvals }); }
    const nextStep = instance.stepInstances.find(value => value.sequence === step.sequence + 1); const operations = [prisma.workflowStepInstance.update({ where: { id: step.id }, data: { status: 'APPROVED', approvals, completedAt: new Date() } }), prisma.workflowAction.create({ data: { instanceId: instance.id, stepInstanceId: step.id, actorUserId: req.auth!.id, actingForUserId, action: 'APPROVE', comment: body.comment } })]; if (nextStep) operations.push(prisma.workflowStepInstance.update({ where: { id: nextStep.id }, data: { status: 'PENDING', startedAt: new Date(), dueAt: nextStep.step.slaHours ? new Date(Date.now() + nextStep.step.slaHours * 3_600_000) : undefined } }) as never); else { operations.push(prisma.workflowInstance.update({ where: { id: instance.id }, data: { status: 'APPROVED', completedAt: new Date() } }) as never); if (instance.subjectType === 'LeaveRequest') operations.push(prisma.leaveRequest.updateMany({ where: { id: instance.subjectId, companyId }, data: { status: 'APPROVED', approvedById: req.auth!.id, approvedAt: new Date() } }) as never); if (instance.subjectType === 'ExpenseClaim') operations.push(prisma.expenseClaim.updateMany({ where: { id: instance.subjectId, companyId }, data: { status: 'APPROVED', approvedById: req.auth!.id, approvedAt: new Date() } }) as never); if(instance.subjectType==='EmployeeServiceRequest')operations.push(prisma.employeeServiceRequest.updateMany({where:{id:instance.subjectId,companyId},data:{status:'APPROVED'}})as never); } await prisma.$transaction(operations); await audit(req, 'APPROVE_WORKFLOW_STEP', `Workflow ${instance.id} step ${step.sequence} approved.`); return ok(res, { status: nextStep ? 'PENDING' : 'APPROVED' });
  } catch (error) { next(error); } });

  router.post('/workflows/delegations', permitted('workflow.review'), async (req: WorkflowRequest, res, next) => { try { const body = z.object({ delegateUserId: uuid, module: z.enum(modules).optional(), startsAt: z.coerce.date(), endsAt: z.coerce.date(), reason: z.string().trim().max(300).optional() }).refine(value => value.endsAt > value.startsAt, 'Delegation end must follow start.').parse(req.body); if (!await prisma.user.findFirst({ where: { id: body.delegateUserId, companyId: req.auth!.companyId } })) return fail(res, 400, 'DELEGATE_INVALID', 'Delegate is not part of this company.'); const value = await prisma.workflowDelegation.create({ data: { ...body, companyId: req.auth!.companyId, delegatorUserId: req.auth!.id } }); await audit(req, 'CREATE_WORKFLOW_DELEGATION', `Approvals delegated to ${body.delegateUserId}.`); return ok(res, value, 201); } catch (error) { next(error); } });
  return router;
}